+++ src/enzyme_ad/jax/Dialects/Comm/CommTypes.cpp
@@ -1,0 +1,7 @@
+#include "src/enzyme_ad/jax/Dialects/Comm/CommDialect.h"
+
+using namespace mlir;
+using namespace mlir::comm;
+
+#define GET_TYPEDEF_CLASSES
+#include "src/enzyme_ad/jax/Dialects/Comm/CommTypes.cpp.inc"
