+++ test/lit_tests/comm_unroll.mlir
@@ -4,10 +4,14 @@ module {
 
   func.func @main(%a : tensor<2x2xf32>) -> tensor<2x2xf32> {
     comm.foo
-    comm.split_branch <[1, 5]> {
-      ^start:
-      comm.foo
-      comm.join
+    comm.split {
+      branch [d1, d4] 
+        ^start:
+        comm.foo
+        comm.join
+      branch [d2]
+        ^start:
+        comm.join
     }
     %start = stablehlo.constant dense<0> : tensor<i32>
